+++ uip/mhshowsbr.c
@@ -469,7 +469,8 @@ raw:
 	}
     }
 
-    if (buflen <= 0 || (ct->c_termproc && buflen <= strlen(ct->c_termproc))) {
+    if (buflen <= 0 ||
+        (ct->c_termproc && (size_t) buflen <= strlen(ct->c_termproc))) {
 	/* content_error would provide a more useful error message
 	 * here, except that if we got overrun, it probably would
 	 * too.
@@ -984,7 +985,8 @@ raw:
 	}
     }
 
-    if (buflen <= 0 || (ct->c_termproc && buflen <= strlen(ct->c_termproc))) {
+    if (buflen <= 0 ||
+        (ct->c_termproc && (size_t) buflen <= strlen(ct->c_termproc))) {
 	/* content_error would provide a more useful error message
 	 * here, except that if we got overrun, it probably would
 	 * too.
